requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
click==8.1.7
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.1.2
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.1.3
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1

requests~=2.32.3
alist3~=1.3.3